+++ changelog.md
@@ -8,6 +8,7 @@
 
 -   [Input] Add new property `fullWidth`.
 -   [Input] Add new property `inputRef`.
+-   [Typography] Add export from index.ts.
 
 ### Changed